start working on final CLI
